Move text highlight to shared
